refactor: move HttpManager to data/manager
