+++ shared/src/commonMain/kotlin/com/example/tasky/agenda/data/AgendaDataSource.kt
@@ -14,8 +14,8 @@ import com.example.tasky.agenda.data.model.TaskPath
 import com.example.tasky.agenda.data.model.UpdateEventBody
 import com.example.tasky.agenda.data.model.UpdateReminderBody
 import com.example.tasky.agenda.data.model.UpdateTaskBody
+import com.example.tasky.common.data.manager.HttpManager
 import com.example.tasky.common.data.util.safeCall
-import com.example.tasky.common.manager.HttpManager
 import com.example.tasky.common.model.BaseError
 import com.example.tasky.common.model.ResultWrapper
 import io.ktor.client.HttpClient
